80769 | Fix PHP Notices on QR Code Check-in

+++ src/Tribe/RSVP.php
@@ -1323,8 +1323,12 @@ class Tribe__Tickets__RSVP extends Tribe__Tickets__Tickets {
 	 * @return void
 	 */
 	public function purge_attendees_transient( $attendee_id ) {
+
 		$event_id = get_post_meta( $attendee_id, self::ATTENDEE_EVENT_KEY, true );
-		Tribe__Post_Transient::instance()->delete( $event_id, Tribe__Tickets__Tickets::ATTENDEES_CACHE );
+
+		if ( $event_id ) {
+			Tribe__Post_Transient::instance()->delete( $event_id, Tribe__Tickets__Tickets::ATTENDEES_CACHE );
+		}
 	}
 
 	/**
